// dm-front/src/components/Settings/UsersApp/UserAppEditModal/types.ts
export type FormState = {
  businessModel: string;
  sellerType: string;
  sellerCategory: string;
  operatingModel: string;
  isBusinessIndividual: boolean;
  isVatRegistered: boolean;
  vatRegisteredName: string;
  vatNumber: string;
  crNumber: string;
  bankName: string;
  registeredAddress: string;
  iban: string;
};